Inclusão de classe de modelo abstrato
